requirements.txt
# Universal Procurement AI - Requirements
# Python 3.10+ required

# OpenAI & LLM
openai>=1.0.0
tiktoken>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2  # Added for Excel export

# Vector Database & Embeddings
chromadb>=0.4.0

# Document Processing & RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
pypdf>=3.17.0
python-docx>=1.0.0
markdown>=3.5.0

# Web UI
streamlit>=1.28.0
plotly>=5.17.0

# Web Framework (API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# Visualization
graphviz>=0.20.1

# PDF Generation
reportlab>=4.0.0

# Logging
loguru>=0.7.0

# Data Validation
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
